requirements.txt
# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Deep learning utilities
tqdm>=4.60.0
matplotlib>=3.4.0
seaborn>=0.11.0
wandb>=0.10.0

# Dataset specific
Pillow>=8.0.0
opencv-python>=4.5.0
albumentations>=1.0.0

# NLP specific (for CivilComments and MultiNLI)
transformers>=4.0.0
nltk>=3.6.0
spacy>=3.0.0

# Medical imaging (for CXR)
pydicom>=2.2.0

# Utilities
pyyaml>=5.4.0
tensorboard>=2.5.0
joblib>=1.0.0

# Development
pytest>=6.2.0
black>=21.5b0
flake8>=3.9.0 